+++ src/server/plugins/engine/models/types.ts
@@ -10,7 +10,7 @@ import {
   type Field
 } from '~/src/server/plugins/engine/components/factory.js'
 import { type RepeatPageController } from '~/src/server/plugins/engine/pageControllers/RepeatPageController.js'
-import { type PageControllerClass } from '~/src/server/plugins/engine/pageControllers/helpers.js'
+import { type PageControllerClass } from '~/src/server/plugins/engine/pageControllers/factory.js'
 import {
   type FormState,
   type FormSubmissionError
