+++ src/server/plugins/engine/pageControllers/helpers.test.ts
@@ -10,10 +10,10 @@ import { FormModel } from '~/src/server/plugins/engine/models/FormModel.js'
 import { PageController } from '~/src/server/plugins/engine/pageControllers/PageController.js'
 import {
   createPage,
-  getProxyUrlForLocalDevelopment,
   isPageController,
   type PageControllerType
-} from '~/src/server/plugins/engine/pageControllers/helpers.js'
+} from '~/src/server/plugins/engine/pageControllers/factory.js'
+import { getProxyUrlForLocalDevelopment } from '~/src/server/plugins/engine/pageControllers/helpers.js'
 import {
   FileUploadPageController,
   QuestionPageController,
